refactor(settings): move Privacy to its own sub-screen
Promote the Privacy section out of PreferencesScreen and into a dedicated
Settings → Account → Privacy sub-screen (third sibling of Profile and
Preferences). Same toggle + Clear-history UI, but on its own screen so it
can grow independently and so the row alignment matches MenuItem rows.

Also fix a pre-existing UX bug surfaced by manual testing: the
permission-denial alert was hardcoded to "Storage Permission Required" /
"App needs access to your storage…" regardless of which permission was
actually denied. Switch the denial branch in requestPermission to use the
generic `permissions.permissionDenied` + new `permissions.appNeedsAccess`
strings so the alert no longer lies about the permission type.

Includes:
- New PrivacyScreen at src/screens/Settings/Privacy/PrivacyScreen.tsx
- ROUTES.SETTINGS_PRIVACY + SCREENS.SETTINGS.PRIVACY + navigator wiring
- Privacy entry added to SettingsScreen's Account menu (Profile →
  Preferences → Privacy)
- i18n keys migrated from preferencesScreen.privacySection.* to a new
  top-level privacyScreen.* namespace (en + cs_cz)
- Privacy rows now use styles.sectionMenuItemTopDescription (the same
  wrapper MenuItem rows use), so the horizontal inset matches the rest
  of the Settings screens
- requestPermission denial alert no longer says "storage" for non-storage
  permissions
