+++ src/libs/Navigation/AppNavigator/ModalStackNavigators/index.tsx
@@ -164,6 +164,9 @@ const SettingsModalStackNavigator =
     [SCREENS.SETTINGS.PREFERENCES.COLOR_PALETTE]: () =>
       require<ReactComponentModule>('@screens/Settings/Preferences/ColorPaletteScreen')
         .default,
+    [SCREENS.SETTINGS.PRIVACY.ROOT]: () =>
+      require<ReactComponentModule>('@screens/Settings/Privacy/PrivacyScreen')
+        .default,
     [SCREENS.SETTINGS.ADMIN.ROOT]: () =>
       require<ReactComponentModule>('@screens/Settings/Admin/AdminScreen')
         .default,
