+++ src/libs/Navigation/types.ts
@@ -120,6 +120,7 @@ type SettingsNavigatorParamList = {
   [SCREENS.SETTINGS.PREFERENCES.UNITS_TO_COLORS]: undefined;
   [SCREENS.SETTINGS.PREFERENCES.DRINKS_TO_UNITS]: undefined;
   [SCREENS.SETTINGS.PREFERENCES.COLOR_PALETTE]: undefined;
+  [SCREENS.SETTINGS.PRIVACY.ROOT]: undefined;
   [SCREENS.SETTINGS.ADMIN.ROOT]: undefined;
   [SCREENS.SETTINGS.ADMIN.FEEDBACK]: undefined;
   [SCREENS.SETTINGS.ADMIN.BUGS]: undefined;
